move gtc2 to gt4py.gtc
