+++ tests/test_unittest/test_gtc/test_gtir_to_pnir.py
@@ -2,9 +2,9 @@ from typing import Iterator
 
 import pytest
 
-from gtc2 import common, gtir
-from gtc2.python import pnir
-from gtc2.python.gtir_to_pnir import GtirToPnir
+from gt4py.gtc import common, gtir
+from gt4py.gtc.python import pnir
+from gt4py.gtc.python.gtir_to_pnir import GtirToPnir
 
 
 @pytest.fixture
